requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Cors
Flask-WTF
python-dotenv
Werkzeug